requirements.txt
streamlit==1.1.0
easyocr==1.4.0
numpy==1.21.2
Pillow==8.3.2
nltk==3.6.5
